add sustain pedal support
Previously, the sampler would not react to the sustain pedal being pressed on the
MIDI controller. This change adds the option to make any notes sustain by pressing
the sustain pedal (MIDI control #64).

Note: this does not affect the existing functionality to simulate sustain without
the pedal by making the note tail longer (via running the app with "--sustain" CLI
argument).

testing: manually tested the sampler by playing it for a while and stressing it in
         the following ways:
         - playing big chords repeatedly with the sustain pedal being pressed -> no
           notes seem to be missed or left playing at the wrong time
         - playing the same notes repeatedly really fast -> click sound is minimal
           and only slightly noticeable when repeating the same note(s) really fast
         - playing this sequence:
            1. play a chord with pedal pressed and lift all but one keys -> all
               notes ringing
            2. release pedal -> only key that is pressed still ringing
            3. press pedal again, then release key -> note still ringing
            4. release pedal with no keys pressed -> silence